requirements.txt
fastapi
uvicorn[standard]
pytesseract
pillow
python-multipart
pydantic
jinja2
gunicorn==23.0.0